changing close error helper to take long ints

+++ 0x15-file_io/holberton.h
@@ -16,7 +16,7 @@ ssize_t read_textfile(const char *filename, size_t letters);
 int create_file(const char *filename, char *text_content);
 int append_text_to_file(const char *filename, char *text_content);
 
-void close_error(int fd);
+void close_error(long int fd);
 void dpf_error(char *readout, char *filename, int exit_c);
 
 #endif /* HOLBERTON_H */
